rdt: split ctrl and mon group into a separate file
Move the code around for better readability and maintainability.
